mobile menu && new color palette
finished mobile nav menu, however, I'm currently giving the main nav bar a major face lift. Will finish that in the morning

I wasn't liking how bland the website was looking so I experimented with new color palettes and decided to keep the preexisting dark green as the secondary color and chose to use a dark blue as the primary color and aliceBlue and a lighter brown for accents and highlights.